mvc패턴으로 변경
컨트롤러에서 비지니스 로직이 있어 서비스로 옮김
MarkerRepositoy Long -> Integer 로 변경

+++ src/main/java/com/today/demo/repository/MarkerRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MarkerRepository extends JpaRepository<Marker, Long> {
+public interface MarkerRepository extends JpaRepository<Marker, Integer> {
     List<Marker> findByVenue(int venue);
 }
